+++ skrl/utils/model_instantiators/torch/shared.py
@@ -14,6 +14,7 @@ from skrl.models.torch import (  # noqa
     MultiCategoricalMixin,
     MultivariateGaussianMixin,
 )
+from skrl.utils.model_instantiators.torch.common import one_hot_encoding  # noqa
 from skrl.utils.model_instantiators.torch.common import convert_deprecated_parameters, generate_containers
 from skrl.utils.spaces.torch import unflatten_tensorized_space  # noqa