+++ skrl/utils/runner/jax/runner.py
@@ -77,6 +77,8 @@ class Runner:
             from skrl.utils.model_instantiators.jax import gaussian_model as component
         elif name == "categoricalmixin":
             from skrl.utils.model_instantiators.jax import categorical_model as component
+        elif name == "multicategoricalmixin":
+            from skrl.utils.model_instantiators.jax import multicategorical_model as component
         elif name == "deterministicmixin":
             from skrl.utils.model_instantiators.jax import deterministic_model as component
         # memory
@@ -319,10 +321,12 @@ class Runner:
             )
             agent_cfg.get("value_preprocessor_kwargs", {}).update({"size": 1, "device": device})
             if agent_cfg.get("exploration", {}).get("noise", None):
+                agent_cfg["exploration"].get("noise_kwargs", {}).update({"device": device})
                 agent_cfg["exploration"]["noise"] = agent_cfg["exploration"]["noise"](
                     **agent_cfg["exploration"].get("noise_kwargs", {})
                 )
             if agent_cfg.get("smooth_regularization_noise", None):
+                agent_cfg.get("smooth_regularization_noise_kwargs", {}).update({"device": device})
                 agent_cfg["smooth_regularization_noise"] = agent_cfg["smooth_regularization_noise"](
                     **agent_cfg.get("smooth_regularization_noise_kwargs", {})
                 )